backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
minio==7.2.3
psycopg2-binary==2.9.9
apscheduler==3.10.4
aiohttp==3.9.1
duckdb==0.9.2
pandas==2.2.0
cryptography==42.0.0
alembic==1.13.1
pymysql==1.1.0
temporalio==1.5.1
croniter==2.0.1
email-validator==2.1.0.post1
boto3==1.34.0
